requirements.txt
astroid==1.4.9
backports.functools-lru-cache==1.3
configparser==3.5.0
Django==1.8.17
django-appconf==1.0.2
django-compressor==2.1
django-debug-toolbar==1.5
django-libsass==0.7
djangorestframework==3.5.3
enum34==1.1.6
flake8==3.2.1
futures==3.0.5
greenlet==0.4.11
gunicorn==19.4.5
isort==4.2.5
lazy-object-proxy==1.2.2
libsass==0.12.3
mccabe==0.5.3
msgpack-python==0.4.8
neovim==0.1.12
olefile==0.44
Pillow==4.0.0
psycopg2==2.6.1
pycodestyle==2.2.0
pyflakes==1.3.0
pylint==1.6.4
python-decouple==3.0
rcssmin==1.0.6
reportlab==3.3.0
requests==2.13.0
rjsmin==1.0.12
six==1.10.0
sqlparse==0.2.2
trollius==2.1
Unipath==1.1
wheel==0.24.0
whitenoise==3.0
wrapt==1.10.8
